refactor(cli): flatten slack command module
Move the Slack command slice to the neutral command root and update its internal exports without changing public CLI or API behavior.
