+++ turbo/apps/cli/src/commands/slack/message/index.ts
@@ -1,7 +1,7 @@
 import { Command } from "commander";
 import { sendCommand } from "./send";
 
-export const zeroSlackMessageCommand = new Command()
+export const slackMessageCommand = new Command()
   .name("message")
   .description("Manage Slack messages")
   .addCommand(sendCommand)
